% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changePassword.R
\name{changePassword}
\alias{changePassword}
\title{Change User Password using Firebase Auth REST API}
\usage{
changePassword(idToken, newPassword)
}
\arguments{
\item{idToken}{A string representing the Firebase ID token of the user.}

\item{newPassword}{A string representing the new password.}
}
\value{
A list containing the response, including the updated tokens if successful.
}
\description{
This function changes the password of a user using the Firebase Auth REST API.
}
\section{Common Error Codes}{

\itemize{
\item INVALID_ID_TOKEN: The provided ID token is invalid.
\item WEAK_PASSWORD: The new password is not strong enough.
\item CREDENTIAL_TOO_OLD_LOGIN_AGAIN: The user needs to sign in again before changing the password.
}
}

\examples{
\dontrun{
id_token <- "user-id-token"
new_password <- "newsecurepassword"
response <- changePassword(id_token, new_password)
print(response)
}
}
