% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConfig.R
\name{create_config}
\alias{create_config}
\title{Create a Firebase Configuration YAML File}
\usage{
create_config(
  api_key,
  projectId,
  authDomain,
  storageBucket,
  databaseURL = NULL,
  appId = NULL,
  output_dir = "."
)
}
\arguments{
\item{api_key}{The Firebase API Key (required).}

\item{projectId}{The Firebase Project ID (required).}

\item{authDomain}{The Firebase Authentication Domain (required).}

\item{storageBucket}{The Firebase Storage Bucket (required).}

\item{databaseURL}{The Firebase Database URL (optional).}

\item{appId}{The Firebase App ID (optional).}

\item{output_dir}{A string specifying the directory where the \code{config.yaml} file will be saved.
Defaults to the current working directory. If the directory does not exist,
it will be created.}
}
\value{
A message indicating the successful creation of the file and environment variables.
}
\description{
This function generates a \code{config.yaml} file with Firebase project settings
and also updates the project's \code{.Renviron} file with the variables.
}
\examples{
\dontrun{
create_config(
  api_key = "example-key",
  projectId = "example-id",
  authDomain = "example-auth",
  storageBucket = "example-bucket"
)
}

}
